fix issues with Int being wrongly used

+++ modules/infra/src/main/scala/scaladex/infra/SqlDatabase.scala
@@ -207,7 +207,7 @@ class SqlDatabase(datasource: HikariDataSource, xa: doobie.Transactor[IO]) exten
     run(ArtifactTable.selectGroupIds.to[Seq])
 
   override def getGroupIds(limit: Int, offset: Int): Future[Seq[Artifact.GroupId]] =
-    run(ArtifactTable.selectGroupIdsPage.to[Seq]((limit, offset)))
+    run(ArtifactTable.selectGroupIdsPage.to[Seq]((limit.toLong, offset.toLong)))
 
   override def getArtifactIds(ref: Project.Reference): Future[Seq[(Artifact.GroupId, Artifact.ArtifactId)]] =
     run(ArtifactTable.selectArtifactIds.to[Seq](ref))
@@ -219,7 +219,7 @@ class SqlDatabase(datasource: HikariDataSource, xa: doobie.Transactor[IO]) exten
     run(ArtifactTable.selectReferencesByGroupId.to[Seq](groupId))
 
   override def getArtifactRefs(groupId: Artifact.GroupId, limit: Int, offset: Int): Future[Seq[Artifact.Reference]] =
-    run(ArtifactTable.selectReferencesByGroupIdPage.to[Seq]((groupId, limit, offset)))
+    run(ArtifactTable.selectReferencesByGroupIdPage.to[Seq]((groupId, limit.toLong, offset.toLong)))
 
   override def insertUser(userId: UUID, userInfo: UserInfo): Future[Unit] =
     run(UserSessionsTable.insert.run((userId, userInfo)).map(_ => ()))
