+++ modules/infra/src/main/scala/scaladex/infra/sql/ArtifactTable.scala
@@ -154,7 +154,7 @@ object ArtifactTable:
     selectRequest(table, Seq("DISTINCT group_id", "artifact_id", "\"version\""), keys = Seq("group_id"))
 
   /** Paged refs for a group: params are (groupId, limit, offset). */
-  val selectReferencesByGroupIdPage: Query[(GroupId, Int, Int), Reference] =
+  val selectReferencesByGroupIdPage: Query[(GroupId, Long, Long), Reference] =
     Query(
       s"""SELECT DISTINCT group_id, artifact_id, "version" FROM $table
          |WHERE group_id = ? ORDER BY artifact_id, "version" LIMIT ? OFFSET ?""".stripMargin
